backend/requirements.txt
# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Environment
python-dotenv>=1.0.0

# File Upload
python-multipart>=0.0.6

# HTTP Client
httpx>=0.25.0

# Supabase
supabase>=2.0.0

# OpenAI
openai>=1.3.0

# File Validation
python-magic>=0.4.27
